+++ packages/fe-fpm-writer/templates/common/EventHandler.ts
@@ -1,5 +1,13 @@
+import ExtensionAPI from 'sap/fe/core/ExtensionAPI';
+import <%- parameters.importType %> from '<%- parameters.importSource %>';
 import MessageToast from 'sap/m/MessageToast';
 
-export function <%- (typeof eventHandlerFnName !== 'undefined' && eventHandlerFnName) || 'onPress' %>() {
+/**
+ * Generated event handler.
+ *
+ * @param this reference to the 'this' that the event handler is bound to.
+ * @param <%- parameters.name %> <%- parameters.description %>
+ */
+export function <%- (typeof eventHandlerFnName !== 'undefined' && eventHandlerFnName) || 'onPress' %>(this: ExtensionAPI, <%- parameters.name %>: <%- parameters.importType %>) {
     MessageToast.show("Custom handler invoked.");
 }