+++ packages/fe-fpm-writer/test/unit/action.test.ts
@@ -289,7 +289,7 @@ describe('CustomAction', () => {
                     line: 8,
                     character: 9
                 },
-                190 + 8 * os.EOL.length
+                196 + 8 * os.EOL.length
             ];
             for (const position of positions) {
                 const testName = `"eventHandler" is object. Append new function to existing js file with position ${