+++ packages/ui5-application-writer/src/data/defaults.ts
@@ -62,6 +62,7 @@ export const enum UI5_DEFAULT {
     OPENUI5_CDN = 'https://openui5.hana.ondemand.com',
     TYPES_VERSION_SINCE = '1.76.0',
     TYPES_VERSION_PREVIOUS = '1.71.18',
+    TYPES_VERSION_BEST = '1.102.7',
     MANIFEST_VERSION = '1.12.0',
     BASE_COMPONENT = 'sap/ui/core/UIComponent'
 }
@@ -110,18 +111,35 @@ export function mergeUi5(ui5: Partial<UI5>): UI5 {
         framework,
         frameworkUrl: ui5.frameworkUrl ?? defaultFrameworkUrl
     };
-    // typesVersion must be a valid npm semantic version, we know they cannot be null as already validated
-    const localSemVer = semVer.valid(semVer.coerce(merged.localVersion))!;
-    const typesVersion = semVer.gte(localSemVer, UI5_DEFAULT.TYPES_VERSION_SINCE)
-        ? localSemVer
-        : UI5_DEFAULT.TYPES_VERSION_PREVIOUS;
 
     merged.descriptorVersion = getManifestVersion(merged.minUI5Version, ui5.descriptorVersion);
-    merged.typesVersion = ui5.typesVersion ?? typesVersion;
+    merged.typesVersion = ui5.typesVersion ?? getTypesVersion(merged.minUI5Version);
     merged.ui5Theme = ui5.ui5Theme ?? 'sap_fiori_3';
     merged.ui5Libs = getUI5Libs(ui5.ui5Libs);
 
     return Object.assign({}, ui5, merged) as UI5;
+}
+
+/**
+ * Get the best types version for the given minUI5Version.
+ * For the latest versions the LTS S/4 on-premise version (1.102.x) is used, for anything before we match the versions as far back as available.
+ *
+ * @param minUI5Version the mininum UI5 version that needs to be supported
+ * @returns semantic version representing the types version.
+ */
+export function getTypesVersion(minUI5Version?: string) {
+    const version = semVer.coerce(minUI5Version);
+    if (!version) {
+        return `~${UI5_DEFAULT.TYPES_VERSION_BEST}`;
+    } else {
+        if (semVer.gte(version, UI5_DEFAULT.TYPES_VERSION_BEST)) {
+            return `~${UI5_DEFAULT.TYPES_VERSION_BEST}`;
+        } else {
+            return semVer.gte(version, UI5_DEFAULT.TYPES_VERSION_SINCE)
+                ? `~${semVer.major(version)}.${semVer.minor(version)}.${semVer.patch(version)}`
+                : UI5_DEFAULT.TYPES_VERSION_PREVIOUS;
+        }
+    }
 }
 
 /**